fixes #1511 - moved save actions to CrudPanel object
