+++ src/app/Http/Controllers/Operations/UpdateOperation.php
@@ -50,7 +50,7 @@ trait UpdateOperation
         // get the info for that entry
         $this->data['entry'] = $this->crud->getEntry($id);
         $this->data['crud'] = $this->crud;
-        $this->data['saveAction'] = $this->getSaveAction();
+        $this->data['saveAction'] = $this->crud->getSaveAction();
         $this->data['fields'] = $this->crud->getUpdateFields($id);
         $this->data['title'] = $this->crud->getTitle() ?? trans('backpack::crud.edit').' '.$this->crud->entity_name;
 
@@ -86,8 +86,8 @@ trait UpdateOperation
         \Alert::success(trans('backpack::crud.update_success'))->flash();
 
         // save the redirect choice for next time
-        $this->setSaveAction();
+        $this->crud->setSaveAction();
 
-        return $this->performSaveAction($item->getKey());
+        return $this->crud->performSaveAction($item->getKey());
     }
 }
